modbus: add protocol extension support
A very much WIP and a proof of concept at this point. The code compiles
with & without the new configuration flag, but is otherwise incomplete.

Currently uses a cumbersome, but simple memory management allocation scheme
where memory is statically allocated based on Kconfig. Currently the user
defined handlers are global, but they should probably be iface specific,
which will result in some memory overhead with static allocation. For very
simple applications it probably doesn't matter and any overhead is
miniscule when compared to the cost of enabling heap support in the kernel.

TODO (at least):
  - unify naming conventions
  - add tests

Signed-off-by: Henrik Lindblom <henrik.lindblom@vaisala.com>